Move DataFrameLike out of the tests
